refactor(powershell): ♻️ merge OutputHelper files
